dollar/requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.3.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
outcome==1.3.0.post0
pydantic==2.3.0
pydantic_core==2.6.3
PySocks==1.7.1
python-slugify==8.0.1
python-telegram-bot==20.6
pytz==2023.3.post1
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
tornado==6.3.3
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzlocal==5.2
urllib3==2.0.7
wsproto==1.2.0
yarl==1.9.3
ymaps==1.3
